requirements.txt
pytz>=2024.1
Pillow>=10.0.0
pycountry>=23.12.0